requirements.txt
streamlit
pandas
google-auth
google-auth-oauthlib
google-api-python-client
requests
python-dotenv
groq
serpapi
logging